🐛[fix]: rebase 후 역할이 중복되는 파일 제거 및 코드 병합
