--- Projects/Core/Networking/API/Sources/API/Attandance/AttandanceAPI.swift
@@ -1,19 +1,0 @@
-//
-//  AttandanceAPI.swift
-//  API
-//
-//  Created by Wonji Suh  on 5/10/25.
-//
-
-import Foundation
-
-public enum AttandanceAPI: String, CaseIterable {
-  case fetchCount
-
-  public var attandanceDescription: String {
-    switch self {
-    case .fetchCount:
-      return "count/"
-    }
-  }
-}
