--- Projects/Core/Networking/Model/Sources/Profile/ProfileModel.swift
@@ -1,29 +1,0 @@
-//
-//  ProfileModel.swift
-//  Model
-//
-//  Created by Wonji Suh  on 5/9/25.
-//
-
-import Foundation
-
-public typealias ProfileModel = BaseResponse<ProfileResponseModel>
-
-public struct ProfileResponseModel: Decodable {
-  let id, name, inviteCodeID, role : String?
-  let team, crew: String?
-  let isStaff: Bool?
-  let createdAt, updatedAt, cohort, responsibility: String?
-  
-  enum CodingKeys: String, CodingKey {
-    case id, name
-    case inviteCodeID = "invite_code_id"
-    case role, team, crew
-    case isStaff = "is_staff"
-    case createdAt = "created_at"
-    case updatedAt = "updated_at"
-    case cohort
-    case responsibility
-  }
-}
-
