+++ Projects/Core/Networking/Service/Sources/QRService/QRService.swift
@@ -14,11 +14,11 @@ import AsyncMoya
 
 public enum QRService {
   case qrAttendanceCheck(code: String)
+  case createQRCode
 }
 
-
-extension QRService : BaseTargetType {
-  public var domain: AttandanceDomain {
+extension QRService: BaseTargetType {
+  public var domain: AttendanceDomain {
     return .qr
   }
   
@@ -26,6 +26,8 @@ extension QRService : BaseTargetType {
     switch self {
     case .qrAttendanceCheck:
       return QRAPI.qrcodeValidate.qrcodeDescription
+    case .createQRCode:
+      return ""
     }
   }
   
@@ -33,22 +35,26 @@ extension QRService : BaseTargetType {
     switch self{
     case .qrAttendanceCheck:
       return .post
+    case .createQRCode:
+      return .post
     }
   }
   
-  public var parameters: [String : Any]? {
+  public var parameters: [String: Any]? {
     switch self {
     case .qrAttendanceCheck(let code):
       let parameters: [String: Any] = ["qr_string": code]
       return parameters
+    case .createQRCode:
+      return nil
     }
   }
   
-  public var headers: [String : String]? {
+  public var headers: [String: String]? {
     return APIHeader.baseHeader
   }
   
-  public var error: [Int : Foundations.NetworkError]? {
+  public var error: [Int: Foundations.NetworkError]? {
     return nil
   }
 }
