+++ Projects/Core/Networking/UseCase/Sources/QRcode/Repository/DefaultQRCodeRepository.swift
@@ -7,6 +7,8 @@
 
 import SwiftUI
 
+import Model
+
 final public class DefaultQRCodeRepository: QRCodeRepositoryProtocol {
   public init() {}
