+++ Projects/Core/Networking/UseCase/Sources/SignUp/Repository/DefaultSignUpRepository.swift
@@ -9,38 +9,20 @@ import Model
 
 final public class DefaultSignUpRepository: SignUpRepositoryProtocol {
   public init() {}
-  
+
   public func registerAccount(
     email: String,
     password: String
   ) async throws -> SignUpDTOModel? {
     return nil
   }
-  
+
   public func validateInviteCode(
     inviteCode: String
   ) async throws -> InviteDTOModel? {
     return nil
   }
-  
-  public func checkEmail(email: String) async throws -> CheckEmailDTO? {
-    return nil
-  }
-  
-  public func registerAccount(
-    userName: String,
-    email: String,
-    password: String
-  ) async throws -> SignUpDTOModel? {
-    return nil
-  }
-  
-  public func validateInviteCode(
-    inviteCode: String
-  ) async throws -> SignUPInviteDTOModel? {
-    return nil
-  }
-  
+
   public func checkEmail(email: String) async throws -> CheckEmailDTO? {
     return nil
   }
